refactoring files&folder structure
