+++ SayHiSemanticUi/graphql/resolvers/bookPost.js
@@ -1,0 +1,14 @@
+const PostBook = require('../../models/PostBook');
+
+module.exports = {
+    Query: {
+        async getBookPosts() {
+          try {
+            const bookPosts = await PostBook.find();
+            return bookPosts;
+          } catch (err) {
+            throw new Error(err);
+          }
+        },
+      },
+}
