+++ SayHiSemanticUi/graphql/resolvers/index.js
@@ -1,0 +1,7 @@
+const bookPostResolvers = require('./bookPost');
+const usersResolvers = require('./users');
+module.exports = {
+  Query: {
+    ...bookPostResolvers.Query,
+  },
+};
