+++ SayHiSemanticUi/graphql/typeDefs.js
@@ -1,0 +1,14 @@
+const {gql} = require('graphql-tag');
+
+module.exports = gql`
+  type PostBook {
+    id:ID!
+    author: String!
+    name: String!
+    username: String!
+    createdAt: String!
+  }
+  type Query {
+    getBookPosts:[PostBook]
+  }
+`;
